a pass at KDoc
